refactor: extract config and parser out of codex-cli.ts
